+++ src/fs/rm.js
@@ -3,6 +3,10 @@ import {onFail} from "../utils/mesLogger.js";
 
 export const rm = async (file) => {
     try {
+        if (!fs.existsSync(file)) {
+            onFail();
+            return;
+        }
         fs.unlinkSync(file);
     }
     catch (e) {
